进行中

+++ details/src/test/java/org/apache/iotdb/api/test/business/TestBlendScenarioAddTS.java
@@ -1,4 +1,4 @@
-package org.apache.iotdb.api.test.bussiness;
+package org.apache.iotdb.api.test.business;
 
 import org.apache.iotdb.api.test.BaseTestSuite;
 import org.apache.iotdb.api.test.utils.CustomDataProvider;
@@ -19,9 +19,12 @@ import org.testng.annotations.Test;
 import java.io.IOException;
 import java.util.*;
 
-public class TestTemplateAlignedAddTS extends BaseTestSuite {
+/**
+ * 需要整理调试
+ */
+public class TestBlendScenarioAddTS extends BaseTestSuite {
     private String database = "root.blendAddTS";
-    private String[] devices = new String[]{database+".alignedUsingTemp", database+".nonAlignedUsingTemp"};
+    private String[] devices = new String[]{database+".alignedUsingTemp", database+".nonAlignedUsingTemp", database+".aligned", database+".nonAligned"};
     private String[] templateNames = new String[]{"aligned_template", "nonAligned_template"};
     private int expectCount = 17;
     private Map<String, Object[]> structureInfo = new LinkedHashMap<>(6);
@@ -34,12 +37,12 @@ public class TestTemplateAlignedAddTS extends BaseTestSuite {
         if (checkStroageGroupExists(database)) {
             session.deleteStorageGroup(database);
         }
+        session.setStorageGroup(database);
         for(String templateName: templateNames) {
             if (checkTemplateExists(templateName)) {
                 session.dropSchemaTemplate(templateName);
             }
         }
-        session.setStorageGroup(database);
         structureInfo.put("s_boolean", new Object[]{TSDataType.BOOLEAN, TSEncoding.RLE, CompressionType.SNAPPY});
         structureInfo.put("s_int", new Object[]{TSDataType.INT32, TSEncoding.RLE, CompressionType.SNAPPY});
         structureInfo.put("s_long", new Object[]{TSDataType.INT64, TSEncoding.RLE, CompressionType.SNAPPY});
@@ -57,10 +60,15 @@ public class TestTemplateAlignedAddTS extends BaseTestSuite {
 
     @AfterClass(enabled = true)
     public void afterClass() throws IoTDBConnectionException, StatementExecutionException {
-        session.deleteStorageGroup(database);
-        if (checkTemplateExists(templateNames[0])) {
-            session.dropSchemaTemplate(templateNames[0]);
+        if (checkStroageGroupExists(database)) {
+            session.deleteStorageGroup(database);
         }
+        for(String templateName: templateNames) {
+            if (checkTemplateExists(templateName)) {
+                session.dropSchemaTemplate(templateName);
+            }
+        }
+
     }
     public Iterator<Object[]> getSingleNormal() throws IOException {
         return new CustomDataProvider().load("data/business-insert-records.csv").getData();
@@ -83,13 +91,13 @@ public class TestTemplateAlignedAddTS extends BaseTestSuite {
         //  IOTDB-5437 StatementExecutionException: 300: COUNT_MEASUREMENTShas not been supported.
 //        assert 6 == session.countMeasurementsInTemplate(templateName) : "查看模版中sensor数目";
         session.setSchemaTemplate(templateNames[index], devices[index]);
-        assert 1+templateCount == getTemplateCount(verbose): "创建模版成功";
-        assert checkTemplateContainPath(templateNames[index], devices[index]) : "挂载模版成功";
+        assert 1+templateCount == getTemplateCount(verbose) : "创建模版成功";
+        assert checkTemplateContainPath(templateNames[index], devices[index]) : "挂载模版成功:"+devices[index];
         assert 1 == getSetPathsCount(templateNames[index], verbose) : "挂载模版成功";
     }
     private void insertTablet(int index, boolean isAligned) throws IOException, IoTDBConnectionException, StatementExecutionException {
-        assert 0 == getTimeSeriesCount(devices[index]+".**", verbose) : "创建前没有符合条件的TimeSeries:"+devices[index];
-        assert 0 == getRecordCount(devices[index], verbose) : "插入前record数目,"+devices[index]+" isAligned="+isAligned;
+        assert 0 == getTimeSeriesCount(devices[index]+".**", true) : "创建前没有符合条件的TimeSeries:"+devices[index];
+        assert 0 == getRecordCount(devices[index], true) : "插入前record数目,"+devices[index]+" isAligned="+isAligned;
         Tablet tablet = new Tablet(devices[index], schemaList, 100);
         int rowIndex = 0;
         int col = 0;
@@ -135,8 +143,8 @@ public class TestTemplateAlignedAddTS extends BaseTestSuite {
             session.insertTablet(tablet);
             Assert.assertThrows(StatementExecutionException.class, () -> session.insertAlignedTablet(tablet));
         }
-        assert 6 == getTimeSeriesCount(devices[index]+".**", verbose) : "成功创建TimeSeries";
-        assert expectCount - 1 == getRecordCount(devices[index], verbose) : "插入record数目,"+devices[index]+" isAligned="+isAligned;
+        assert 6 == getTimeSeriesCount(devices[index]+".**", true) : "成功创建TimeSeries";
+        assert expectCount - 1 == getRecordCount(devices[index], true) : "插入record数目,"+devices[index]+" isAligned="+isAligned;
     }
     private void doUpdate(int index, boolean isAligned) throws IoTDBConnectionException, StatementExecutionException {
         long timestamp = 1669109398772L;
@@ -174,10 +182,6 @@ public class TestTemplateAlignedAddTS extends BaseTestSuite {
     private void doUpdateAfterAddTS(int index, boolean isAligned) throws IoTDBConnectionException, StatementExecutionException {
         long timestamp1 = 1669109508000L;
         long timestamp2 = 1669109398772L;
-        System.out.println("###### 前  ######");
-        getCount("select count(*) from "+devices[index]+" where time="+timestamp1+";", verbose);
-        getCount("select count(*) from "+devices[index]+" where time="+timestamp2+";", verbose);
-        System.out.println("###### 前  ######");
         List<Long> times = new ArrayList<>(2);
         List<List<String>> valueList = new ArrayList<>(2);
         List<List<String>> measurementList = new ArrayList<>(2);
@@ -202,89 +206,10 @@ public class TestTemplateAlignedAddTS extends BaseTestSuite {
         } else {
             session.insertStringRecordsOfOneDevice(devices[index], times, measurementList, valueList);
         }
-        System.out.println("###### 后  ######");
-        getCount("select count(*) from "+devices[index]+" where time="+timestamp1+";", verbose);
-        getCount("select count(*) from "+devices[index]+" where time="+timestamp2+";", verbose);
-        System.out.println("###### 后  ######");
         checkQueryResult("select s_float from "+devices[index]+" where time="+timestamp1+";", 13.33);
         checkQueryResult("select appendFloat from "+devices[index]+" where time="+timestamp1+";", 34567.0);
         checkQueryResult("select s_long from "+devices[index]+" where time="+timestamp2+";", timestamp2);
         checkQueryResult("select appendFloat from "+devices[index]+" where time="+timestamp2+";", 69134.0);
-    }
-
-    private void doUpdateAfterAddTS2(int index, boolean isAligned) throws IoTDBConnectionException, StatementExecutionException {
-        long timestamp1 = 1669109508000L;
-        long timestamp2 = 1669109398772L;
-        System.out.println("###### 前  ######");
-        getCount("select count(*) from "+devices[index]+" where time="+timestamp1+";", verbose);
-        getCount("select count(*) from "+devices[index]+" where time="+timestamp2+";", verbose);
-        System.out.println("###### 前  ######");
-        List<Long> times = new ArrayList<>(2);
-        List<List<Object>> valueList = new ArrayList<>(2);
-        List<List<String>> measurementList = new ArrayList<>(2);
-        times.add(timestamp1);
-        times.add(timestamp2);
-        measurementList.add(measurements);
-        measurementList.add(measurements);
-
-        for (int i=0; i<2; i++) {
-            List<Object> values = new ArrayList<>(7);
-            values.add(true);
-            values.add(i);
-            values.add(times.get(i));
-            values.add((i+1)*13.33f);
-            values.add((i+1)*144.44);
-            values.add("insertAlignedRecord:"+i);
-            values.add((i+1)*34567f);
-            valueList.add(values);
-            if (isAligned) {
-                session.insertAlignedRecord(devices[index], times.get(i), measurements, dataTypes, valueList.get(i));
-            } else {
-                session.insertRecord(devices[index], times.get(i), measurements, dataTypes, valueList.get(i));
-            }
-        }
-        System.out.println("###### 后  ######");
-        getCount("select count(*) from "+devices[index]+" where time="+timestamp1+";", verbose);
-        getCount("select count(*) from "+devices[index]+" where time="+timestamp2+";", verbose);
-        System.out.println("###### 后  ######");
-        checkQueryResult("select s_float from "+devices[index]+" where time="+timestamp1+";", 13.33);
-        checkQueryResult("select appendFloat from "+devices[index]+" where time="+timestamp1+";", 34567.0);
-        checkQueryResult("select s_long from "+devices[index]+" where time="+timestamp2+";", timestamp2);
-        checkQueryResult("select appendFloat from "+devices[index]+" where time="+timestamp2+";", 69134.0);
-    }
-    private void doUpdateAfterAddTS_origin(int index, boolean isAligned) throws IoTDBConnectionException, StatementExecutionException {
-        Long[] times = new Long[]{1669109509000L, 1669109398772L};
-        System.out.println("###### 前  ######");
-        getCount("select count(*) from "+devices[index]+" where time="+times[0]+";", verbose);
-        getCount("select count(*) from "+devices[index]+" where time="+times[1]+";", verbose);
-        System.out.println("###### 前  ######");
-        Tablet tablet = new Tablet(devices[index], schemaList, 100);
-        int rowIndex = 0;
-        tablet.initBitMaps();
-
-        for (int i=0; i<2; i++) {
-            rowIndex = tablet.rowSize++;
-            tablet.addTimestamp(rowIndex, times[i]);
-            tablet.addValue(schemaList.get(0).getMeasurementId(), rowIndex, true);
-            tablet.addValue(schemaList.get(1).getMeasurementId(), rowIndex, i);
-            tablet.addValue(schemaList.get(2).getMeasurementId(), rowIndex, times[i]);
-            tablet.addValue(schemaList.get(3).getMeasurementId(), rowIndex, (i+1)*130.33f);
-            tablet.addValue(schemaList.get(4).getMeasurementId(), rowIndex, (i+1)*14400.44);
-            tablet.addValue(schemaList.get(5).getMeasurementId(), rowIndex, "doUpdateAfterAddTS_origin:"+i);
-        }
-        if (isAligned) {
-            session.insertAlignedTablet(tablet);
-        } else {
-            session.insertTablet(tablet);
-        }
-        System.out.println("###### 后  ######");
-        getCount("select count(*) from "+devices[index]+" where time="+times[0]+";", verbose);
-        getCount("select count(*) from "+devices[index]+" where time="+times[1]+";", verbose);
-        System.out.println("###### 后  ######");
-        checkQueryResult("select s_float from "+devices[index]+" where time="+times[0]+";", 130.33);
-        checkQueryResult("select s_text from "+devices[index]+" where time="+times[0]+";", "doUpdateAfterAddTS_origin:0");
-        checkQueryResult("select s_long from "+devices[index]+" where time="+times[1]+";", times[1]);
-        checkQueryResult("select s_text from "+devices[index]+" where time="+times[1]+";", "doUpdateAfterAddTS_origin:1");
     }
 
     @Test(priority = 10)
@@ -292,56 +217,97 @@ public class TestTemplateAlignedAddTS extends BaseTestSuite {
         createTemplate(0, true);
     }
     @Test(priority = 20)
+    public void testCreateNonAlignedTemplate() throws IoTDBConnectionException, IOException, StatementExecutionException {
+        createTemplate(1, false);
+    }
+    @Test(priority = 30)
+    public void testInsertAlignedTS() throws IoTDBConnectionException, IOException, StatementExecutionException {
+        insertTablet(2, true);
+    }
+    @Test(priority = 40)
     public void testInsertAlignedTemplateTS() throws IoTDBConnectionException, IOException, StatementExecutionException {
         insertTablet(0, true);
     }
-    @Test(priority = 30)
-    public void testUpdateAlignedTS() throws IoTDBConnectionException, StatementExecutionException {
+    @Test(priority = 50)
+    public void testInsertNonAlignedTemplateTS() throws IoTDBConnectionException, IOException, StatementExecutionException {
+        insertTablet(1, false);
+    }
+    @Test(priority = 60)
+    public void testInsertNonAlignedTS() throws IoTDBConnectionException, IOException, StatementExecutionException {
+        insertTablet(3, false);
+    }
+    @Test(priority = 70)
+    public void testUpdateAlignedTemplateTS() throws IoTDBConnectionException, StatementExecutionException {
         doUpdate(0, true);
     }
-    @Test(priority = 40)
-    public void testAddTS() throws IoTDBConnectionException, StatementExecutionException {
-        // 追加创建一个TS
-        List<String> measurement = new ArrayList<>(1);
-        List<TSDataType> dataType = new ArrayList<>(1);
-        List<TSEncoding> encoding = new ArrayList<>(1);
-        List<CompressionType> compressor = new ArrayList<>(1);
-        List<String> alias = new ArrayList<>(1);
-        measurement.add("appendFloat");
-        dataType.add(TSDataType.FLOAT);
-        encoding.add(TSEncoding.GORILLA);
-        compressor.add(CompressionType.UNCOMPRESSED);
-        alias.add("append_float");
-        Assert.assertThrows(StatementExecutionException.class, ()->{
-            session.createAlignedTimeseries(devices[0], measurement, dataType, encoding, compressor, alias);
-        });
-        Assert.assertThrows(StatementExecutionException.class, ()->{
-            session.createAlignedTimeseries(devices[0]+".child", measurement, dataType, encoding, compressor, alias);
-        });
-        assert 6 == getTimeSeriesCount(devices[0]+".**", verbose) : "追加一个普通的TS到模版节点失败";
-        session.createAlignedTimeseries(database+".d_normal", measurement, dataType, encoding, compressor, alias);
-
+    @Test(priority = 80)
+    public void testUpdateNonAlignedTemplateTS() throws IoTDBConnectionException, StatementExecutionException {
+        doUpdate(1, false);
     }
-    @Test(priority = 50)
-    public void testUpdateAfterAddTS_alignedTemp() throws IoTDBConnectionException, StatementExecutionException {
-        // 追加TS结构
+    @Test(priority = 90)
+    public void testUpdateAlignedTS() throws IoTDBConnectionException, StatementExecutionException {
+        doUpdate(2, true);
+    }
+    @Test(priority = 100)
+    public void testUpdateNonAlignedTS() throws IoTDBConnectionException, IOException, StatementExecutionException {
+        doUpdate(3, false);
+    }
+    @Test(priority = 101)
+    public void testUpdateAfterAddTS_aligned() throws IoTDBConnectionException, StatementExecutionException {
         measurements.add("appendFloat");
         dataTypes.add(TSDataType.FLOAT);
-
-//      // 使用 insertAlignedStringRecordsOfOneDevice
-        // IOTDB-5441
+        doUpdateAfterAddTS(2, true);
+    }
+    @Test(priority = 102)
+    public void testUpdateAfterAddTS_nonAligned() throws IoTDBConnectionException, StatementExecutionException {
+        doUpdateAfterAddTS(3, false);
+    }
+    @Test(priority = 103)
+    public void testUpdateAfterAddTS_alignedTemp() throws IoTDBConnectionException, StatementExecutionException {
+//        List<String> measurement = new ArrayList<>(1);
+//        List<TSDataType> dataType = new ArrayList<>(1);
+//        List<TSEncoding> encoding = new ArrayList<>(1);
+//        List<CompressionType> compressor = new ArrayList<>(1);
+//        List<String> alias = new ArrayList<>(1);
+//        measurement.add("appendFloat");
+//        dataType.add(TSDataType.FLOAT);
+//        encoding.add(TSEncoding.GORILLA);
+//        compressor.add(CompressionType.UNCOMPRESSED);
+//        alias.add("append_float");
+//        session.createAlignedTimeseries(devices[0], measurement, dataType, encoding, compressor, alias);
+//        assert 7 == getTimeSeriesCount(devices[0]+".**", true) : "追加一个TS";
         doUpdateAfterAddTS(0, true);
-
-        // 使用 insertAlignedRecord
-//        doUpdateAfterAddTS2(0, true);
-
+    }
+    @Test(priority = 104)
+    public void testUpdateAfterAddTS_nonAlignedTemp() throws IoTDBConnectionException, StatementExecutionException {
+        doUpdateAfterAddTS(1, false);
     }
 
-    @Test(priority = 60)
-    public void testInsertWithoutAdd() throws IoTDBConnectionException, StatementExecutionException {
-        //使用tablet
-//        schemaList.add(new MeasurementSchema("appendFloat", TSDataType.FLOAT, TSEncoding.GORILLA, CompressionType.SNAPPY));
-        doUpdateAfterAddTS_origin(0, true);
+    @Test(priority = 110)
+    public void testDropTS() throws IoTDBConnectionException, StatementExecutionException {
+        for(int index=2; index<4; index++) {
+            assert 7 == getTimeSeriesCount(devices[index] + ".*", verbose) : "删除前TimeSeries为6:"+devices[index];
+            session.deleteTimeseries(devices[index] + ".*");
+            assert 0 == getTimeSeriesCount(devices[index] + ".*", verbose) : "删除后TimeSeries为0:"+devices[index];
+        }
+    }
+    @Test(priority = 120)
+    public void testDropTemplate() throws IoTDBConnectionException, StatementExecutionException, IOException {
+        for (int i=0; i<templateNames.length; i++) {
+            String templateName = templateNames[i];
+            assert checkTemplateExists(templateName) : "模版存在";
+            assert checkTemplateContainPath(templateName, devices[i]) : "挂载模版路径";
+            deactiveTemplate(templateName, devices[i]);
+            session.unsetSchemaTemplate(devices[i], templateName);
+            assert false == checkTemplateContainPath(templateName, devices[i]) : "解除模版路径成功";
+
+            // IOTDB-5436 StatementExecutionException: 300: Modify template has not been supported.
+//        session.deleteNodeInTemplate(templateName, loadNode);
+            assert 0 == getActivePathsCount(templateName, verbose) : "没有已激活路径";
+            assert 0 == getSetPathsCount(templateName, verbose) : "没有已挂载路径";
+            session.dropSchemaTemplate(templateName);
+            assert false == checkTemplateExists(templateName) : "删除模版成功";
+        }
 
     }
